+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 2.0.1-SNAPSHOT
 - Adds `ngram` function
 - Adds `is_empty` and `is_nonempty` functions
+- Provide a raw view of the query in the RequestResult
 
 ## 2.0.0 (March 19, 2018)
 - Adds support for @query type (user defined functions)