+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+## 2.0.2-SNAPSHOT
+
 ## 2.0.1 (March 24, 2018)
 - Adds `ngram` function
 - Adds `is_empty` and `is_nonempty` functions